Add mutation for removing user from location

+++ src/main/kotlin/com/ledahl/apps/recieppyapi/model/Location.kt
@@ -4,6 +4,6 @@ data class Location(
         val id: Long = 0,
         val name: String,
         val address: String?,
-        val owner: Int,
+        val owner: Long,
         val inviteCode: String
 )
